ANGLE EGL and GLES libraries should link with lib dl. https://bugs.webkit.org/show_bug.cgi?id=226920
Patch by Eleni Maria Stea <estea@igalia.com> on 2021-06-11
Reviewed by Kenneth Russell.

GLES and EGL that are generated from ANGLE should link
with libdl because they use dlsym and dlopen. Also,
with the updated CMakeLists.txt we can safely link with libEGL
when USE_ANGLE_WEBGL is set because the symbols of EGL are now
found.

* CMakeLists.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278768 268f45cc-cd09-0410-ab3c-d52691b4dbfc