[iOS] Permit access to 'platform-name' https://bugs.webkit.org/show_bug.cgi?id=227032 <rdar://problem/79334360>
Reviewed by Per Arne Vollan.

Telemetry shows that some media code needs access to 'platform-name'. We allow this in the
WebContent process, but didn't include it in the GPU Process.

Make the WebContent and GPU Process consistent in this access. Also adopt the standard rule
for this used in the system sandbox, which restricts the access to the IOKit class family that
has this property.

* Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb:
* Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb.in:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278881 268f45cc-cd09-0410-ab3c-d52691b4dbfc