FileReader.readAsArrayBuffer misses checking for the type of blob param https://bugs.webkit.org/show_bug.cgi?id=226640 <rdar://problem/79193198>
Reviewed by Youenn Fablet.

Source/WebCore:

The Blob parameter to FileReader.readAsArrayBuffer / readAsBinaryString / readAsText / readAsDataURL
should not be nullable as per the specification:
- https://w3c.github.io/FileAPI/#APIASynch

This patch aligns us with the specification and with the behavior of both Blink and Gecko.

Test: fast/files/filereader-invalid-blob.html

* fileapi/FileReader.cpp:
(WebCore::FileReader::readAsArrayBuffer):
(WebCore::FileReader::readAsBinaryString):
(WebCore::FileReader::readAsText):
(WebCore::FileReader::readAsDataURL):
* fileapi/FileReader.h:
* fileapi/FileReader.idl:

LayoutTests:

Add layout test coverage. I have verified that this test is passing in Firefox and Chrome as well.

* fast/files/filereader-invalid-blob-expected.txt: Added.
* fast/files/filereader-invalid-blob.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278951 268f45cc-cd09-0410-ab3c-d52691b4dbfc