Clear state after process crash so that we know to re-issue sandbox extensions https://bugs.webkit.org/show_bug.cgi?id=227103 <rdar://problem/77588760>
Reviewed by Per Arne Vollan.

We dynamically extend access to certain mach services when the ATTACHMENT_ELEMENT feature
is enabled. If a process with this feature activated crashes and WebKit relaunches it, we
don't grant the new process the same extensions because we believe they were already
previously sent.

This patch clears the state after a process crash so that it properly handles the restart case.

* UIProcess/WebProcessProxy.cpp:
(WebKit::WebProcessProxy::shutDown): Clear state.
(WebKit::WebProcessProxy::processDidTerminateOrFailedToLaunch): Clear state.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279014 268f45cc-cd09-0410-ab3c-d52691b4dbfc