[GLIB] REGRESSION(r278701): numerous WebRTC tests failing after libwebrtc update to M92 https://bugs.webkit.org/show_bug.cgi?id=227172
Patch by Philippe Normand <pnormand@igalia.com> on 2021-06-21
Reviewed by Adrian Perez de Castro.

Non-apple ports do not use the bundled libvpx library, they rely on the host libvpx instead.
LibWebRTC expects to depend on the bundled version, thus using its API. So we need to be
very careful with which libvpx version is used from the host, in order to avoid API/ABI
issues. Hence the version check should specify a version that is API/ABI compatible with
what libwebrtc expects.

This patch also re-enables data channel support, which was disabled during the update to M92
in CMake ports.

* CMakeLists.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279065 268f45cc-cd09-0410-ab3c-d52691b4dbfc