AppHighlights disappear on page reload https://bugs.webkit.org/show_bug.cgi?id=227141 rdar://78190331
Reviewed by Devin Rousso.

Source/WebCore:

Preserve the state of highlight visibility so that is correct on page-reload.

* dom/Document.cpp:
(WebCore::Document::appHighlightRegister):
* page/ChromeClient.h:
(WebCore::ChromeClient::appHighlightsVisiblility const):

Source/WebKit:

When a page is reloaded, a new set of highlights is loaded, but the state of visibility of
the highlights is not preserved.
We need to preserve the state of highlights so they are correctly shown on reload.

* Shared/WebPageCreationParameters.cpp:
(WebKit::WebPageCreationParameters::encode const):
(WebKit::WebPageCreationParameters::decode):
* Shared/WebPageCreationParameters.h:
* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::appHighlightsVisiblility const):
* WebProcess/WebCoreSupport/WebChromeClient.h:
* WebProcess/WebPage/WebPage.cpp:
(WebKit::m_appHighlightsVisible):
(WebKit::WebPage::setAppHighlightsVisibility):
(WebKit::m_lastNavigationWasAppBound): Deleted.
* WebProcess/WebPage/WebPage.h:
(WebKit::WebPage::appHighlightsVisiblility const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279078 268f45cc-cd09-0410-ab3c-d52691b4dbfc