WebKit_iosmac build failure: SystemPreviewSPI.h:29:9: fatal error: 'AssetViewer/ARQuickLookWebKitItem.h' file not found https://bugs.webkit.org/show_bug.cgi?id=227507 <rdar://problem/79902043>
Reviewed by Dean Jackson.

When we added new SPIs in Source/WebCore/PAL/pal/spi/ios/SystemPreviewSPI.h as part of r279312 we also included this file
from a new file that is compiled on Catalyst. This revealed an issue where existing SPIs included in that file weren't
actually available on Catalyst. We now make sure we compile those out on Catalyst.

Source/WebCore/PAL:

* pal/spi/ios/SystemPreviewSPI.h:

Source/WTF:

* wtf/PlatformHave.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279392 268f45cc-cd09-0410-ab3c-d52691b4dbfc