[JSC] Stop generating default parameter code if class constructor is called without 'new' https://bugs.webkit.org/show_bug.cgi?id=227547 rdar://78821453
Reviewed by Mark Lam.

JSTests:

* stress/calling-non-callable-constructors.js: Added.
(shouldThrow):

Source/JavaScriptCore:

We already do not generate body bytecode when class constructor is called without 'new' because many features including "super()" assume
that they generate bytecode only when it is called as a constructor. But we are not doing that for default parameters' bytecode generation.
This patch stops generating bytecode for default parameters if class constructor is called without 'new'.

* bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::generate):
(JSC::BytecodeGenerator::BytecodeGenerator):
* runtime/ConstructorKind.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279448 268f45cc-cd09-0410-ab3c-d52691b4dbfc